Issue #1147
Implemented FC_UNUSED, FC_TODO in specific way for GCC (which does not support #pragma unused as clang does).
Other warning messages eliminated.

+++ libraries/fc/src/crypto/elliptic_common.cpp
@@ -85,8 +85,8 @@ namespace fc { namespace ecc {
             ssl_bignum order;
             FC_ASSERT( EC_GROUP_get_order( group, order, ctx ) );
             private_key_secret bin;
-            FC_ASSERT( BN_num_bytes( order ) == bin.data_size() );
-            FC_ASSERT( BN_bn2bin( order, (unsigned char*) bin.data() ) == bin.data_size() );
+            FC_ASSERT( static_cast<size_t>(BN_num_bytes( order )) == bin.data_size() );
+            FC_ASSERT( static_cast<size_t>(BN_bn2bin( order, (unsigned char*) bin.data() )) == bin.data_size() );
             return bin;
         }
 
@@ -104,8 +104,8 @@ namespace fc { namespace ecc {
             FC_ASSERT( EC_GROUP_get_order( group, order, ctx ) );
             BN_rshift1( order, order );
             private_key_secret bin;
-            FC_ASSERT( BN_num_bytes( order ) == bin.data_size() );
-            FC_ASSERT( BN_bn2bin( order, (unsigned char*) bin.data() ) == bin.data_size() );
+            FC_ASSERT( static_cast<size_t>(BN_num_bytes( order )) == bin.data_size() );
+            FC_ASSERT( static_cast<size_t>(BN_bn2bin( order, (unsigned char*) bin.data() )) == bin.data_size() );
             return bin;
         }
